Design a Simple Automaton (Finite State Machine) solution
https://www.codewars.com/kata/design-a-simple-automaton-finite-state-machine/python

Create a finite automaton that has three states. Finite automatons are
the same as finite state machines for our purposes.

Our simple automaton, accepts the language of A, defined as {0, 1} and
should have three states,
q1, q2, and q3.

q1 is our start state. We begin reading commands from here.
q2 is our accept state. We return true if this is our last state.

q1 moves to q2 when given a 1, and stays at q1 when given a 0.
q2 moves to q3 when given a 0, and stays at q2 when given a 1.
q3 moves to q2 when given a 0 or 1.

Our automaton should return whether we end in our accepted state, or not
(true/false.)